spring-data-jpa/src/main/java/com/tommy/datajpa/member/domain/MemberJpaRepository.java
package com.tommy.datajpa.member.domain;

import org.springframework.stereotype.Repository;

import javax.persistence.EntityManager;
import javax.persistence.PersistenceContext;

@Repository
public class MemberJpaRepository {

    @PersistenceContext
    private EntityManager entityManager;

    public Member save(Member member) {
        entityManager.persist(member);
        return member;
    }

    public Member findById(Long id) {
        return entityManager.find(Member.class, id);
    }
}
